Moved and renamed classes and packages
